Lazy property bug fix
re-declaring the lazy property is not always working depending on when in the startup chain the call to get the session store is used.  This can be before the lazy property is re-declared.  However this can be directly set on the singleton property and this results in proper functionality.  However we loose lazy loading.  This however is not a problem because the session store will definitely always be required, so the lazy loading is not providing anything for us.