Fix javascript code of smile_web_impex
Using a callback to evaluate whether the button should be hidden
does not always work. The callback is not guaranteed to be called instantaneously.

For example, when refreshing the screen, the export button disapears.

Instead, use the session to store flags to indicate whether the user is member of
the import/export groups.